fix: split trackedSpawn deadline into startup/teardown, make it configurable (#1583)
TRACKED_SPAWN_DEADLINE_MS (10s) was shared by teardown ops (sessions
close/stop/cancel) and startup ops (sessions ensure, via
createSession/loadSession/applyReasoningEffort). `sessions ensure`
measured a real-world median of 8.15s, so under concurrency nax was
SIGTERMing healthy acpx processes and silently falling back to
context-less sessions.

- Split into trackedSpawnDeadlineMs (teardown, 10s) and
  trackedSpawnStartupDeadlineMs (startup, 30s), both threaded through
  SpawnAcpClient/SpawnAcpSession, the ACP adapter, AgentManager, and
  SessionManager, and bound to config.agent.acp in the Zod schema.
- loadSession now logs when a `sessions ensure` deadline-timeout
  (exitCode -1) drives the context-less fallback, instead of treating
  it identically to a genuine session miss.
- Split spawn-client.ts (SpawnAcpSession -> spawn-client-session.ts,
  shared deps -> spawn-client-deps.ts) and adapter.ts
  (closePhysicalSession -> adapter-close-physical.ts) to stay under
  the file-size ratchet.
