+++ src/agents/acp/adapter-close-physical.ts
@@ -1,0 +1,36 @@
+/**
+ * AcpAgentAdapter.closePhysicalSession implementation — split out of adapter.ts
+ * to stay under the file-size ratchet.
+ */
+
+import { getSafeLogger } from "@/logger";
+import { _acpAdapterDeps } from "./adapter-lifecycle";
+
+export async function closePhysicalSession(
+  agentName: string,
+  handle: string,
+  workdir: string,
+  options?: { force?: boolean; signal?: AbortSignal },
+): Promise<void> {
+  const cmdStr = `acpx ${agentName}`;
+  const client = _acpAdapterDeps.createClient(cmdStr, workdir, undefined, undefined);
+  try {
+    await client.start();
+    try {
+      if (client.closeSession) {
+        await client.closeSession(handle, agentName, options?.signal);
+        // AC-83: hard-terminate (acpx stop) when force=true, e.g. for errored sessions
+        if (options?.force) {
+          await (client as { forceStop?: (agentName: string) => Promise<void> }).forceStop?.(agentName).catch(() => {});
+        }
+      } else if (client.loadSession) {
+        const session = await client.loadSession(handle, agentName, "approve-reads");
+        if (session) await session.close({ forceTerminate: options?.force, signal: options?.signal }).catch(() => {});
+      }
+    } catch (err) {
+      getSafeLogger()?.warn("acp-adapter", `[close] Failed to close session ${handle}`, { error: String(err) });
+    }
+  } finally {
+    await client.close().catch(() => {});
+  }
+}
