+++ src/config/index.ts
@@ -105,6 +105,7 @@ export {
   mutationCheckConfigSelector,
   executionGatesConfigSelector,
   finishConfigSelector,
+  trackedSpawnDeadlines,
 } from "./selectors";
 export { createConfigLoader } from "./loader-runtime";
 export type { ConfigLoader } from "./loader-runtime";
